refactor: move constants to helper folder
