+++ src/helpers/constRegistration.ts
@@ -1,0 +1,19 @@
+export enum InputTypes {
+  NAME1 = 'signin-name',
+  NAME2 = 'login-name',
+  PASSPASSWORD1 = 'signin-password',
+  PASSPASSWORD2 = 'login-password',
+  EMAIL1 = 'signin-email',
+  EMAIL2 = 'login-email',
+}
+
+export enum FormTypes {
+  LOGIN = 'login',
+  SIGNIN = 'sign-in',
+}
+
+export enum FetchStatus {
+  START = 'start',
+  DONE = 'done',
+  INIT = 'init',
+}
