+++ src/requests/userWords/createUserWordAPI.ts
@@ -1,5 +1,5 @@
 import { IUserWord } from 'src/requests/interfaceAPI';
-import { Methods, UrlPath, Headers } from 'src/requests/constantsAPI';
+import { Methods, UrlPath, Headers } from 'src/helpers/constRequestsAPI';
 
 const createUserWordAPI = async (userId: string, wordId: string, wordData: IUserWord) => {
   try {
