+++ src/requests/userWords/getAllUserWordsAPI.ts
@@ -1,5 +1,5 @@
 import { IUserWord } from 'src/requests/interfaceAPI';
-import { Methods, UrlPath, Headers } from 'src/requests/constantsAPI';
+import { Methods, UrlPath, Headers } from 'src/helpers/constRequestsAPI';
 
 const getAllUserWordsAPI = async (userId: string) => {
   try {
