+++ src/requests/userWords/getUserWordAPI.ts
@@ -1,5 +1,5 @@
 import { IUserWord } from 'src/requests/interfaceAPI';
-import { Methods, UrlPath, Headers } from 'src/requests/constantsAPI';
+import { Methods, UrlPath, Headers } from 'src/helpers/constRequestsAPI';
 
 const getUserWordAPI = async (userId: string, wordId: string) => {
   try {
