+++ src/requests/users/getUserTokenAPI.ts
@@ -1,5 +1,5 @@
 import { ISignInResponse } from 'src/requests/interfaceAPI';
-import { Methods, UrlPath, Headers } from 'src/requests/constantsAPI';
+import { Methods, UrlPath, Headers } from 'src/helpers/constRequestsAPI';
 
 const getUserAPI = async (id: string) => {
   try {
